Rename account_auth file
