+++ src/middleware/authentication.ts
@@ -5,7 +5,7 @@ import { CustomerService } from '../services/customer.js'
 import * as dotenv from 'dotenv'
 import { withLogto, handleAuthRoutes, LogtoExpressConfig } from '@logto/express'
 import { configLogToExpress } from '../types/constants.js'
-import { AccountAuthHandler } from './auth/account_auth.js'
+import { AccountAuthHandler } from './auth/account-auth.js'
 import { CredentialAuthHandler } from './auth/credential_auth.js'
 import { DidAuthHandler } from './auth/did_auth.js'
 import { KeyAuthHandler } from './auth/key_auth.js'
